requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
pillow>=9.5.0
tqdm>=4.65.0
opencv-python>=4.7.0
pandas>=2.0.0
scikit-image>=0.20.0
selenium>=4.10.0
requests>=2.31.0
webdriver-manager>=3.8.6
beautifulsoup4>=4.12.0 